pyproject.toml
[project]
name = "ChNu-Practics-2024"
description = "Empowering Innovation: Bridging Startups and Investors for Ukraine's Economic Growth"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "django>=4.2,<5.0",
  "djangorestframework",
  "django-cors-headers",
  "psycopg2-binary",
  "djangorestframework-simplejwt",
  "python-dotenv",
  "gunicorn",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
development = ["pre-commit", "ruff", "mypy", "django-stubs[compatible-mypy]"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = 'migrations/*'

[tool.ruff]
src = ["src"]
line-length = 88
ignore = [
  "E721", # Do not compare types, use `isinstance()`
  # pydocstyle ignore
  "D100", # Missing docstring in public module
  "D101", # Missing docstring in public class
  "D102", # Missing docstring in public method
  "D103", # Missing docstring in public function
  "D104", # Missing docstring in public package
  "D105", # Missing docstring in magic method
  "D106", # Missing docstring in public nested class
  "D107", # Missing docstring in `__init__`
  "D209", # [*] Multi-line docstring closing quotes should be on a separate line
  "D205", # 1 blank line required between summary line and description
  "D400", # First line should end with a period
  "D401", # First line of docstring should be in imperative mood
  "D402", # First line should not be the function's signature
  "D404", # First word of the docstring should not be "This"
  "D415", # First line should end with a period, question mark, or exclamation point
  "E501", # Line too long
]
select = ["B", "D", "E", "F", "I", "N", "Q", "W"]

[tool.ruff.lint.isort]
force-single-line = true
combine-as-imports = true
force-sort-within-sections = true
order-by-type = false
lines-after-imports = 2

[tool.mypy]
follow_imports = "silent"
ignore_missing_imports = true

# optionals
no_implicit_optional = true

# strictness
allow_untyped_globals = false
allow_redefinition = false
implicit_reexport = false
strict_equality = true

# warnings
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_redundant_casts = true
warn_unreachable = true

# error output
show_column_numbers = true
show_error_context = true
show_error_codes = true
pretty = true
color_output = true
error_summary = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"
